Fix pagination issue with BatchDeleter
Prior to this change, the pagination behavior of BatchDeleter was to
window over 10 records at a time, incrementing the `start` parameter by
one each time. The issue there was that we were making 10 requests to
each record in the entire collection (ramping up using the first 10
records). The pagination was working like this:

0, 1, 2, 3, 4, 5, 6, 7, 8, 9
1, 2, 3, 4, 5, 6, 7, 8, 9, 10
2, 3, 4, 5, 6, 7, 8, 9, 10, 11

and so on. With this change, we're correctly paginating through by
increasing the `start` offset by the (now configurable) `batch_size`.