Use persistent connection when batching requests to CDM
When running the BatchDeleteWorker, we'll end up making several batches
of requests to CONTENTdm's OAI endpoint. Each batch is handled by its
own Sidekiq worker, but for the duration of that job, we can utilize a
persistent connection to CONTENTdm rather than making a new one for each
request. This will save time and bandwidth as this job runs.